requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-cors-headers==2.1.0
djangorestframework==3.7.7
idna==2.6
pytz==2018.3
requests==2.18.4
urllib3==1.22
